moved the "make valid" at the beginning, to have valid polygons right at the start Fix the test to have the skeleton lines output
